adding middlewares post link to redux tutorials
This is tooting my own horn but no pride invested so shoot this down if it isn't a good fit. But I've used your list as a resource (thank you!), and missing some material focusing on middlewares on the list is what prompted me to write the post in the first place! :)